acesstoken api auth

+++ service/tasksImpl.js
@@ -1,12 +1,15 @@
 const TaskSchema = require("../models/PlanModel");
 const { ObjectId } = require("mongodb");
-const handleError = require("../utils/error");
-
+const handleError = require("../utils/extend");
 
 const getTasks = async (req, res) => {
   try {
-    const findTask = await TaskSchema.find({});
-    res.status(200).json(findTask);
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if(userId) {
+      const findTask = await TaskSchema.find({ userId });
+      res.status(200).json(findTask);
+    }
   } catch (error) {
     handleError(error, res);
   }
@@ -14,8 +17,24 @@ const getTasks = async (req, res) => {
 
 const addTasks = async (req, res) => {
   try {
-    const newTask = await TaskSchema.create(req.body);
-    res.status(200).json(newTask);
+    const { title, projectName, description, priority, category, createdAt } =
+      req.body;
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if (userId) {
+      const task = {
+        userId: userId,
+        title: title,
+        projectName: projectName,
+        description: description,
+        priority: priority,
+        category: category,
+        createdAt: createdAt,
+      };
+
+      const newTask = await TaskSchema.create(task);
+      res.status(200).json(newTask);
+    }
   } catch (error) {
     handleError(error, res);
   }
@@ -23,10 +42,25 @@ const addTasks = async (req, res) => {
 
 const updateTask = async (req, res) => {
   try {
+    const { title, projectName, description, priority, category, createdAt } =
+      req.body;
     const id = req.params;
-    await TaskSchema.findByIdAndUpdate(new ObjectId(id), req.body);
-    const updatedTask = await TaskSchema.findById(new ObjectId(id));
-    res.status(200).json(updatedTask);
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if (userId) {
+      const task = {
+        userId: userId,
+        title: title,
+        projectName: projectName,
+        description: description,
+        priority: priority,
+        category: category,
+        createdAt: createdAt,
+      };
+      await TaskSchema.findByIdAndUpdate(new ObjectId(id), task);
+      const updatedTask = await TaskSchema.findById(new ObjectId(id));
+      res.status(200).json(updatedTask);
+    }
   } catch (error) {
     handleError(error, res);
   }
@@ -34,20 +68,27 @@ const updateTask = async (req, res) => {
 
 const getTaskById = async (req, res) => {
   try {
-    const id = req.params;
-    const getTask = await TaskSchema.findById(new ObjectId(id));
-    res.status(200).json(getTask);
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if (userId) {
+      const id = req.params;
+      const getTask = await TaskSchema.findById(new ObjectId(id));
+      res.status(200).json(getTask);
+    }
   } catch (error) {
     handleError(error, res);
   }
 };
 
-
 const deleteTasks = async (req, res) => {
   try {
-    const id = req.params;
-    const deleteTask = await TaskSchema.findByIdAndDelete(new ObjectId(id));
-    res.status(200).json(deleteTask);
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if(userId) {
+      const id = req.params;
+      const deleteTask = await TaskSchema.findByIdAndDelete(new ObjectId(id));
+      res.status(200).json(deleteTask);
+    }
   } catch (error) {
     handleError(error, res);
   }
@@ -55,11 +96,24 @@ const deleteTasks = async (req, res) => {
 
 const deleteResolvedTasks = async (req, res) => {
   try {
-    const deleteResolved = await TaskSchema.deleteMany({category: "RESOLVED"})
-    res.status(200).json(deleteResolved);
+    const accessToken = req.headers.authorization;
+    const userId = validateAndGetUserIdFromAccessToken(accessToken);
+    if(userId){
+      const deleteResolved = await TaskSchema.deleteMany({
+        category: "RESOLVED",
+      });
+      res.status(200).json(deleteResolved);
+    }
   } catch (error) {
     handleError(error, res);
   }
-}
+};
 
-module.exports = {getTasks, addTasks, updateTask, getTaskById, deleteTasks, deleteResolvedTasks}
+module.exports = {
+  getTasks,
+  addTasks,
+  updateTask,
+  getTaskById,
+  deleteTasks,
+  deleteResolvedTasks,
+};
